+++ actix-service/src/apply.rs
@@ -1,218 +1,172 @@
-use futures::{try_ready, Async, Future, IntoFuture, Poll};
+use std::marker::PhantomData;
 
-use super::{FnNewTransform, FnTransform};
-use super::{
-    IntoNewService, IntoNewTransform, IntoService, IntoTransform, NewService, NewTransform,
-    Service, Transform,
-};
+use futures::{Async, Future, IntoFuture, Poll};
+
+use super::{IntoNewService, IntoService, NewService, Service};
 
 /// `Apply` service combinator
-pub struct Apply<T, S>
+pub struct Apply<T, F, In, Out>
 where
-    T: Transform<S>,
-    T::Error: From<S::Error>,
-    S: Service,
+    T: Service,
 {
-    transform: T,
-    service: S,
+    service: T,
+    f: F,
+    r: PhantomData<(In, Out)>,
 }
 
-impl<T, S> Apply<T, S>
+impl<T, F, In, Out> Apply<T, F, In, Out>
 where
-    T: Transform<S>,
-    T::Error: From<S::Error>,
-    S: Service,
-{
-    /// Create new `Apply` combinator
-    pub fn new<T1: IntoTransform<T, S>, S1: IntoService<S>>(
-        transform: T1,
-        service: S1,
-    ) -> Self {
-        Self {
-            transform: transform.into_transform(),
-            service: service.into_service(),
-        }
-    }
-}
-
-impl<F, S, Req, Out> Apply<FnTransform<F, S, Req, Out>, S>
-where
-    F: FnMut(Req, &mut S) -> Out,
+    T: Service,
+    F: FnMut(In, &mut T) -> Out,
     Out: IntoFuture,
-    Out::Error: From<S::Error>,
-    S: Service,
+    Out::Error: From<T::Error>,
 {
     /// Create new `Apply` combinator
-    pub fn new_fn<S1: IntoService<S>>(service: S1, transform: F) -> Self {
+    pub fn new<I: IntoService<T>>(service: I, f: F) -> Self {
         Self {
             service: service.into_service(),
-            transform: transform.into_transform(),
+            f,
+            r: PhantomData,
         }
     }
 }
 
-impl<T, S> Clone for Apply<T, S>
+impl<T, F, In, Out> Clone for Apply<T, F, In, Out>
 where
-    S: Service + Clone,
-    T::Error: From<S::Error>,
-    T: Transform<S> + Clone,
+    T: Service + Clone,
+    F: Clone,
 {
     fn clone(&self) -> Self {
         Apply {
             service: self.service.clone(),
-            transform: self.transform.clone(),
+            f: self.f.clone(),
+            r: PhantomData,
         }
     }
 }
 
-impl<T, S> Service for Apply<T, S>
+impl<T, F, In, Out> Service for Apply<T, F, In, Out>
 where
-    T: Transform<S>,
-    T::Error: From<S::Error>,
-    S: Service,
+    T: Service,
+    F: FnMut(In, &mut T) -> Out,
+    Out: IntoFuture,
+    Out::Error: From<T::Error>,
 {
-    type Request = T::Request;
-    type Response = T::Response;
-    type Error = T::Error;
-    type Future = T::Future;
+    type Request = In;
+    type Response = Out::Item;
+    type Error = Out::Error;
+    type Future = Out::Future;
 
     fn poll_ready(&mut self) -> Poll<(), Self::Error> {
-        try_ready!(self.service.poll_ready());
-        self.transform.poll_ready()
+        self.service.poll_ready().map_err(|e| e.into())
     }
 
-    fn call(&mut self, req: Self::Request) -> Self::Future {
-        self.transform.call(req, &mut self.service).into_future()
+    fn call(&mut self, req: In) -> Self::Future {
+        (self.f)(req, &mut self.service).into_future()
     }
 }
 
 /// `ApplyNewService` new service combinator
-pub struct ApplyNewService<T, S, C>
+pub struct ApplyNewService<T, F, In, Out, Cfg>
 where
-    T: NewTransform<S::Service, C, InitError = S::InitError>,
-    T::Error: From<S::Error>,
-    S: NewService<C>,
+    T: NewService<Cfg>,
 {
-    transform: T,
-    service: S,
-    _t: std::marker::PhantomData<C>,
+    service: T,
+    f: F,
+    r: PhantomData<(In, Out, Cfg)>,
 }
 
-impl<T, S, C> ApplyNewService<T, S, C>
+impl<T, F, In, Out, Cfg> ApplyNewService<T, F, In, Out, Cfg>
 where
-    T: NewTransform<S::Service, C, InitError = S::InitError>,
-    T::Error: From<S::Error>,
-    S: NewService<C>,
+    T: NewService<Cfg>,
+    F: FnMut(In, &mut T::Service) -> Out + Clone,
+    Out: IntoFuture,
+    Out::Error: From<T::Error>,
 {
     /// Create new `ApplyNewService` new service instance
-    pub fn new<T1: IntoNewTransform<T, S::Service, C>, S1: IntoNewService<S, C>>(
-        transform: T1,
-        service: S1,
-    ) -> Self {
+    pub fn new<F1: IntoNewService<T, Cfg>>(service: F1, f: F) -> Self {
         Self {
-            transform: transform.into_new_transform(),
+            f,
             service: service.into_new_service(),
-            _t: std::marker::PhantomData,
+            r: PhantomData,
         }
     }
 }
 
-impl<F, S, In, Out, Cfg>
-    ApplyNewService<FnNewTransform<F, S::Service, In, Out, S::InitError, Cfg>, S, Cfg>
+impl<T, F, In, Out, Cfg> Clone for ApplyNewService<T, F, In, Out, Cfg>
 where
-    F: FnMut(In, &mut S::Service) -> Out + Clone,
+    T: NewService<Cfg> + Clone,
+    F: FnMut(In, &mut T::Service) -> Out + Clone,
     Out: IntoFuture,
-    Out::Error: From<S::Error>,
-    S: NewService<Cfg>,
-{
-    /// Create new `Apply` combinator factory
-    pub fn new_fn<S1: IntoNewService<S, Cfg>>(service: S1, transform: F) -> Self {
-        Self {
-            service: service.into_new_service(),
-            transform: FnNewTransform::new(transform),
-            _t: std::marker::PhantomData,
-        }
-    }
-}
-
-impl<T, S, C> Clone for ApplyNewService<T, S, C>
-where
-    T: NewTransform<S::Service, C, InitError = S::InitError> + Clone,
-    T::Error: From<S::Error>,
-    S: NewService<C> + Clone,
 {
     fn clone(&self) -> Self {
         Self {
             service: self.service.clone(),
-            transform: self.transform.clone(),
-            _t: std::marker::PhantomData,
+            f: self.f.clone(),
+            r: PhantomData,
         }
     }
 }
 
-impl<T, S, C> NewService<C> for ApplyNewService<T, S, C>
+impl<T, F, In, Out, Cfg> NewService<Cfg> for ApplyNewService<T, F, In, Out, Cfg>
 where
-    T: NewTransform<S::Service, C, InitError = S::InitError>,
-    T::Error: From<S::Error>,
-    S: NewService<C>,
+    T: NewService<Cfg>,
+    F: FnMut(In, &mut T::Service) -> Out + Clone,
+    Out: IntoFuture,
+    Out::Error: From<T::Error>,
 {
-    type Request = T::Request;
-    type Response = T::Response;
-    type Error = T::Error;
-    type Service = Apply<T::Transform, S::Service>;
+    type Request = In;
+    type Response = Out::Item;
+    type Error = Out::Error;
+    type Service = Apply<T::Service, F, In, Out>;
 
     type InitError = T::InitError;
-    type Future = ApplyNewServiceFuture<T, S, C>;
+    type Future = ApplyNewServiceFuture<T, F, In, Out, Cfg>;
 
-    fn new_service(&self, cfg: &C) -> Self::Future {
+    fn new_service(&self, cfg: &Cfg) -> Self::Future {
+        ApplyNewServiceFuture::new(self.service.new_service(cfg), self.f.clone())
+    }
+}
+
+pub struct ApplyNewServiceFuture<T, F, In, Out, Cfg>
+where
+    T: NewService<Cfg>,
+    F: FnMut(In, &mut T::Service) -> Out + Clone,
+    Out: IntoFuture,
+{
+    fut: T::Future,
+    f: Option<F>,
+    r: PhantomData<(In, Out)>,
+}
+
+impl<T, F, In, Out, Cfg> ApplyNewServiceFuture<T, F, In, Out, Cfg>
+where
+    T: NewService<Cfg>,
+    F: FnMut(In, &mut T::Service) -> Out + Clone,
+    Out: IntoFuture,
+{
+    fn new(fut: T::Future, f: F) -> Self {
         ApplyNewServiceFuture {
-            fut_t: self.transform.new_transform(cfg),
-            fut_s: self.service.new_service(cfg),
-            service: None,
-            transform: None,
+            f: Some(f),
+            fut,
+            r: PhantomData,
         }
     }
 }
 
-pub struct ApplyNewServiceFuture<T, S, C>
+impl<T, F, In, Out, Cfg> Future for ApplyNewServiceFuture<T, F, In, Out, Cfg>
 where
-    T: NewTransform<S::Service, C, InitError = S::InitError>,
-    T::Error: From<S::Error>,
-    S: NewService<C>,
+    T: NewService<Cfg>,
+    F: FnMut(In, &mut T::Service) -> Out + Clone,
+    Out: IntoFuture,
+    Out::Error: From<T::Error>,
 {
-    fut_s: S::Future,
-    fut_t: T::Future,
-    service: Option<S::Service>,
-    transform: Option<T::Transform>,
-}
-
-impl<T, S, C> Future for ApplyNewServiceFuture<T, S, C>
-where
-    T: NewTransform<S::Service, C, InitError = S::InitError>,
-    T::Error: From<S::Error>,
-    S: NewService<C>,
-{
-    type Item = Apply<T::Transform, S::Service>;
+    type Item = Apply<T::Service, F, In, Out>;
     type Error = T::InitError;
 
     fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
-        if self.transform.is_none() {
-            if let Async::Ready(transform) = self.fut_t.poll()? {
-                self.transform = Some(transform);
-            }
-        }
-
-        if self.service.is_none() {
-            if let Async::Ready(service) = self.fut_s.poll()? {
-                self.service = Some(service);
-            }
-        }
-
-        if self.transform.is_some() && self.service.is_some() {
-            Ok(Async::Ready(Apply {
-                service: self.service.take().unwrap(),
-                transform: self.transform.take().unwrap(),
-            }))
+        if let Async::Ready(service) = self.fut.poll()? {
+            Ok(Async::Ready(Apply::new(service, self.f.take().unwrap())))
         } else {
             Ok(Async::NotReady)
         }
@@ -225,7 +179,7 @@ mod tests {
     use futures::{Async, Future, Poll};
 
     use super::*;
-    use crate::{NewService, Service};
+    use crate::{IntoService, NewService, Service, ServiceExt};
 
     #[derive(Clone)]
     struct Srv;
@@ -245,10 +199,14 @@ mod tests {
     }
 
     #[test]
-    fn test_apply() {
-        let mut srv = Apply::new_fn(Srv, |req: &'static str, srv| {
-            srv.call(()).map(move |res| (req, res))
-        });
+    fn test_call() {
+        let blank = |req| Ok(req);
+
+        let mut srv = blank
+            .into_service()
+            .apply_fn(Srv, |req: &'static str, srv| {
+                srv.call(()).map(move |res| (req, res))
+            });
         assert!(srv.poll_ready().is_ok());
         let res = srv.call("srv").poll();
         assert!(res.is_ok());
@@ -258,22 +216,6 @@ mod tests {
     #[test]
     fn test_new_service() {
         let new_srv = ApplyNewService::new(
-            |req: &'static str, srv: &mut Srv| srv.call(()).map(move |res| (req, res)),
-            || Ok::<_, ()>(Srv),
-        );
-        if let Async::Ready(mut srv) = new_srv.new_service(&()).poll().unwrap() {
-            assert!(srv.poll_ready().is_ok());
-            let res = srv.call("srv").poll();
-            assert!(res.is_ok());
-            assert_eq!(res.unwrap(), Async::Ready(("srv", ())));
-        } else {
-            panic!()
-        }
-    }
-
-    #[test]
-    fn test_new_service_fn() {
-        let new_srv = ApplyNewService::new_fn(
             || Ok::<_, ()>(Srv),
             |req: &'static str, srv| srv.call(()).map(move |res| (req, res)),
         );